market-research-platform/news-fetcher/requirements.txt
# filepath: market-research-platform/news-fetcher/requirements.txt
# Pin versions for reproducible builds.

# --- Web framework --------------------------------------------------------
fastapi==0.111.0
uvicorn[standard]==0.29.0

# --- LlamaIndex core ------------------------------------------------------
llama-index-core==0.10.40
llama-index-vector-stores-postgres==0.1.14

# --- LLM providers (configurable via LLM_PROVIDER env var) ----------------
llama-index-llms-gemini==0.1.11
llama-index-embeddings-gemini==0.1.8
llama-index-llms-groq==0.1.4
llama-index-llms-openai==0.1.22
llama-index-embeddings-openai==0.1.10
llama-index-embeddings-huggingface==0.2.2
google-generativeai>=0.5.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0

# --- Database --------------------------------------------------------------
sqlalchemy==2.0.30
asyncpg==0.29.0
pgvector==0.2.5

# --- News fetching ---------------------------------------------------------
httpx==0.27.0
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.2.2

# --- Task scheduling -------------------------------------------------------
apscheduler==3.10.4

# --- Utilities -------------------------------------------------------------
pydantic==2.7.1
pydantic-settings==2.2.1
